mergify: Remove approved review when a PR changes
When people have reviewed a PR, and the contributor updates it, the
reviews are staying valid. This is unfortunate as the PR might need new
corrections.

Ideally a positive review (Approved) needs to be given again. And there
needs to be a confirmation that a negative review (Changes requested)
has been addressed.

Closes: #505
Signed-off-by: Niels de Vos <ndevos@redhat.com>