cephfs: Choose kernel mounter when storage class mounter is kernel
Currently, the kernel mounter is chosen only when the kernel version is >= 4.17,
as the quota feature required for PV creation, is not available in older kernels.
With this patch we are changing the behaviour to choose kernel mounter if the
mounter specified in SC is kernel, irrespective of the kernel version.

Fixes #617

Signed-off-by: Poornima G <pgurusid@redhat.com>