Use a wrapper for formatAndMount with rbd specific actions.
Currently rbd CSI plugin uses formatAndMount of
mount.SafeFormatAndMount. This does not allow to pass or use
specific formatting arguments with it. This patch introduce
RBD specific formatting options with both `xfs` and `ext4`,
for example: `-E no-discard` with ext4 and `-k` option with
XFS to boost formatting performance of RBD device.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>