[Experimental] Add RBD mirroring support based on image snapshots
This commit is fully experimental, with a hack/slash approach to get
RBD mirroring working.

The RBD mirroring mode of snapshot based mirroring is enabled in
this commit, using under development Ceph Octopus release.

The basic design is to name the images with the pvc name-namespace
as its header, and to trigger seraching for such an image when the
CreateVolume parameters specify "mirrored", instead of first creating
the the image.

Signed-off-by: ShyamsundarR <srangana@redhat.com>