test: test functionality of travis with github
Github considers all the checks as passed even in the case when there is an error
in travis.yml and ci doesn't run. This results in auto merging of the code without a valid
CI run, trying to repro the same by introducing a parsing error in travis.yml.
Issue: https://github.com/ceph/ceph-csi/issues/1011

Signed-off-by: Mudit Agarwal <muagarwa@redhat.com>