ci: Enable stale bot to auto close the old issues
atm, the daysUntilStale is 60 which is really huge and it
will take some time to trigger the bot itself and come into
action. This PR reduce or make stale bot active on cases (7days)
to reduce the huge backlog and set daysUntilClose to 25 days for now.
It gives enough window for issues to be revisited.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>