cephfs: fix regression in subvolumegroups
If the PVC is created by ceph-csi v2.0.0 the subvolume
will be default ie "csi" if the user updates the cephcsi
to v3.0.0 and have the subvolume group in configmap,
remounting of older PVC will start failing as the ceph-csi
will read the subvolumegroup from configmap and use it for
mounting the PVC created cephcsi <3.0.0.

This commit fixes it by adding subvolumegroupname in
the volumecontext and use it to get rootpath of subvolume.

Signed-off-by: Madhu Rajanna <madhupr007@gmail.com>