rbd: add vault token authentication support
Fixes: https://github.com/ceph/ceph-csi/issues/1500

Solution: The vault token auth can use token based auth in a
setup. If we want to use vault token based authentication in
a setup, the ceph-csi-encryption-kms-config should have below
values set:

`
       "vaultTokenAuth": "true",
       "vaultTokenFromSecret": "vault-token",
`

Here the first param `vaultTokenAuth` say whether token based
auth is enabled or not. If set, the vault auth token will be fetched
from node stage/provisioner secret of the key `vault-token`.
ie the secret key is the value/variable set for `vaultTokenFromSecret`
parameter in the configmap.

Signed-off-by: Humble Chirammal <hchiramm@redhat.com>